Re #12 -
This introduces riviera-expanded-context-variables which keeps track of
any variables that have been expanded by variable name.

We use the tree-widget-after-toggle-functions hook to process the expand
or collapsing of a widget through riviera-tree-widget-notify which then
adds or removes the proper variable name from the list.

On initialization of the context display, in
riviera-session-context-init we add the :open property which checks if
the variable name is a member of the expanded list.